listo

+++ Tarea 4 - Ventaneo/Procesos/helper.h
@@ -1,6 +1,6 @@
 #ifndef HELPER_H
 #define HELPER_H
-int* reservarMemoria();
+double* reservarMemoria();
 void llenarArreglo(double* arreglo);
 void imprimirArreglo(double* arreglo);
 #endif
